Added second control type
Two control types are now implemented and accessible through a FAB. Also boundries for all triggers were set so they wouln't go out of their background circle. "getData()" method was added to both control types which returns a array with speeds between -1,1. It depends on the control type how many variables are in the array